# build(wasm): optimise in-game build time