--- a/Cargo.toml
+++ b/Cargo.toml
@@ -7,3 +7,6 @@ members = [
     "src/systems/systems_wasm",
     "src/systems/a320_hydraulic_simulation_graphs",
 ]
+
+[profile.release]
+lto = true