## Summary of Changes
Optimises the WASM artefacts by using Link Time Optimisation (LTO) and wasm-opt. This increases the `build.sh` build time a bit, but provides significant gains on in-sim compilation time. 

On my machine the time spent on in-game WASM compilation when all three modules needed compilation went down from 364 seconds to 185 seconds, effectively cutting compilation time in half. Thus, as a result our users, testers and devs will spend less time waiting for MSFS WASM compilation after they installed or built a newer version of the FBW A32NX. 

This change should also improve running performance, though this wasn't measured (and in general the impact on performance of these modules is deemed to be relatively low).

### wasm-opt optimisation levels
wasm-opt supports multiple levels of optimisation, where higher levels of optimisation lead to slower compilation time during development, but can lead to shorter compilation time of the WASM within MSFS.

`systems.wasm` is optimised at level 3, as there were no problems in doing so.
`fbw.wasm` and `fadec.wasm` are optimised at level 1, as any higher optimisation level lead to the following error being emitted during compilation in MSFS:
> Exception c000001d

### Artefact size

| Artefact     | Before | LTO   | LTO + wasm-opt |
|--------------|--------|-------|----------------|
| systems.wasm | 2129 KiB | *     | 558 KiB (level 3)          |
| fbw.wasm     | 1449 KiB | 975 KiB | 630 KiB (level 1)          |
| fadec.wasm   | 482 KiB  | 463 KiB | 249 KiB (level 1)          |

\* not measured

### MSFS build time

| Artefact     | Before | LTO | LTO + wasm-opt |
|--------------|--------|-----|----------------|
| systems.wasm | 211 s   | *   | 70 s (level 3)            |
| fbw.wasm     | 114 s   | 95 s | 83 s (level 1)          |
| fadec.wasm   | 39 s    | 39 s | 32 s (level 1)          |

\* not measured

## Additional context
<!-- Add any other context about the pull request here. -->

<!-- You may optionally provide your discord username, so that we may contact you directly about the issue. -->
Discord username (if different from GitHub): SjotgunSjonnie#9623

## Testing instructions
Ensure systems, fbw and the fadec still work by checking a few of their features. I suggest doing a full flight to detect any issues that might occur over time, as these kinds of changes can sometimes cause the weirdest bugs 😀.

<!-- DO NOT DELETE THIS -->
## How to download the PR for QA

Every new commit to this PR will cause a new A32NX artifact to be created, built, and uploaded.

1. Make sure you are signed in to GitHub
1. Click on the **Checks** tab on the PR
1. On the left side, click on the bottom **PR** tab
1. Click on the **A32NX** download link at the bottom of the page