@frankkopp 2021-06-23
Quality Assurance Tester Report

Discord: Cdr_Maverick#6475
Object of testing: #5188
Tier of Testing: 1
Date: 23.6.2021

Testing Process: 
Full flight C&D to shutdown (simbrief planned route)

Negatives: None

Testing Results: Passed 
For normal flight / did not test any specific features 
I suspect adding optimization should not change functionality as long as the compiler does its work correctly. 

Conclusions: Probably more test flights should be done but looks good so far.